Historical Daily Values of Delphi Research Groups Combined Indicator for Indiana

Processed data for the Indiana COVIDcast Dashboard that shows historic daily values of Delphi Research Group's Combined Indicator for each methropolitan statistical area.

dash_ci_line

Format

An object of class tbl_df (inherits from tbl, data.frame) with 4425 rows and 3 columns.

Details

Data Dictionary entry

Examples

head(dash_ci_line)

#> # A tibble: 6 x 3
#>   name        time_value value
#>   <chr>       <date>     <dbl>
#> 1 Bloomington 2020-04-15  1.27
#> 2 Bloomington 2020-04-16  1.25
#> 3 Bloomington 2020-04-17  1.3 
#> 4 Bloomington 2020-04-18  1.57
#> 5 Bloomington 2020-04-19  1.35
#> 6 Bloomington 2020-04-20  1.71