Historical Google Maps Mobility Indices for Indiana

Processed data for the Google Maps Mobility Indices line chart that shows percent differences from a mobility baseline that was calculated during the 5-week period Jan 3 – Feb 6, 2020. These percent changes show how lengths-of-stay at different locations have changed from pre-pandemic conditions.

goog_mob_ind

Format

An object of class tbl_df (inherits from tbl, data.frame) with 1680 rows and 4 columns.

Details

Data Dictionary entry

Examples

head(goog_mob_ind)

#> # A tibble: 6 x 4
#>   date       activity               index weekend
#>   <date>     <chr>                  <dbl> <lgl>  
#> 1 2020-02-15 Retail and Recreation  0.095 TRUE   
#> 2 2020-02-15 Workplace              0.01  TRUE   
#> 3 2020-02-15 Residential           -0.01  TRUE   
#> 4 2020-02-16 Retail and Recreation  0.08  TRUE   
#> 5 2020-02-16 Workplace              0.01  TRUE   
#> 6 2020-02-16 Residential           -0.01  TRUE