Historical Daily Cases and Deaths by Race for Indiana

Raw data that shows daily case and death counts by race.

ind_race_complete

Format

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 4160 rows and 58 columns.

Details

Data Dictionary entry

Examples

head(ind_race_complete)

#> # A tibble: 6 x 58
#>   date       state cases_total cases_white cases_black cases_latin_x cases_asian
#>   <date>     <chr>       <dbl>       <dbl>       <dbl>         <dbl>       <dbl>
#> 1 2020-06-17 IN          41013       18254        5234            NA         650
#> 2 2020-06-21 IN          42423       18786        5317            NA         673
#> 3 2020-06-24 IN          43140       19422        5399            NA         678
#> 4 2020-06-28 IN          44930       20045        5513            NA         692
#> 5 2020-07-01 IN          45952       20773        5637            NA         709
#> 6 2020-07-05 IN          48008       21638        5769            NA         718
#> # ... with 51 more variables: cases_aian <lgl>, cases_nhpi <lgl>,
#> #   cases_multiracial <lgl>, cases_other <dbl>, cases_unknown <dbl>,
#> #   cases_ethnicity_hispanic <dbl>, cases_ethnicity_non_hispanic <dbl>,
#> #   cases_ethnicity_unknown <dbl>, deaths_total <dbl>, deaths_white <dbl>,
#> #   deaths_black <dbl>, deaths_latin_x <dbl>, deaths_asian <dbl>,
#> #   deaths_aian <lgl>, deaths_nhpi <lgl>, deaths_multiracial <lgl>,
#> #   deaths_other <dbl>, deaths_unknown <dbl>, deaths_ethnicity_hispanic <dbl>,
#> #   deaths_ethnicity_non_hispanic <dbl>, deaths_ethnicity_unknown <dbl>,
#> #   hosp_total <lgl>, hosp_white <lgl>, hosp_black <lgl>, hosp_latin_x <lgl>,
#> #   hosp_asian <lgl>, hosp_aian <lgl>, hosp_nhpi <lgl>, hosp_multiracial <lgl>,
#> #   hosp_other <lgl>, hosp_unknown <lgl>, hosp_ethnicity_hispanic <lgl>,
#> #   hosp_ethnicity_non_hispanic <lgl>, hosp_ethnicity_unknown <lgl>,
#> #   tests_total <lgl>, tests_white <lgl>, tests_black <lgl>,
#> #   tests_latin_x <lgl>, tests_asian <lgl>, tests_aian <dbl>, tests_nhpi <dbl>,
#> #   tests_multiracial <dbl>, tests_other <lgl>, tests_unknown <lgl>,
#> #   tests_ethnicity_hispanic <lgl>, tests_ethnicity_non_hispanic <lgl>,
#> #   tests_ethnicity_unknown <dbl>, cases_latinx <lgl>, deaths_latinx <lgl>,
#> #   hosp_latinx <lgl>, tests_latinx <dbl>