Historical Weekly Positivity Rates and Daily Cases per 100,000 Indiana Residents

Processed data for the Indiana COVIDcast Dashboard that shows weekly positivity rate and daily cases per 100,000 residents per metropolitan statistical area.

msa_cases100_posrate_historic

Format

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 7530 rows and 7 columns.

Details

Data Dictionary entry

Examples

head(msa_cases100_posrate_historic)

#> # A tibble: 6 × 7
#>   date       msa         geo_value cases_100k start_date pos_rate data_date 
#>   <date>     <chr>           <dbl>      <dbl> <date>        <dbl> <date>    
#> 1 2020-04-02 Bloomington     14020       2.95 NA           NA     2021-10-08
#> 2 2020-04-03 Bloomington     14020       3.38 NA           NA     2021-10-08
#> 3 2020-04-04 Bloomington     14020       2.95 2020-03-29    0.109 2021-10-08
#> 4 2020-04-05 Bloomington     14020       2.70 NA           NA     2021-10-08
#> 5 2020-04-06 Bloomington     14020       2.53 NA           NA     2021-10-08
#> 6 2020-04-07 Bloomington     14020       2.53 NA           NA     2021-10-08